+++ robotbase/src/main/java/com/team2357/frc2023/commands/intake/IntakeArmDeployCommand.java
@@ -4,8 +4,8 @@ import com.team2357.frc2023.subsystems.IntakeArmSubsystem;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class DeployIntakeCommand extends CommandBase {
-    public DeployIntakeCommand() {
+public class IntakeArmDeployCommand extends CommandBase {
+    public IntakeArmDeployCommand() {
         addRequirements(IntakeArmSubsystem.getInstance());
     }
