+++ robotbase/src/main/java/com/team2357/frc2023/commands/intake/IntakeHandoffCommandGroup.java
@@ -7,10 +7,10 @@ import com.team2357.frc2023.commands.scoring.util.RotateArmToPositionCommand;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 
-public class IntakeHandoffCommand extends SequentialCommandGroup {
-    public IntakeHandoffCommand() {
+public class IntakeHandoffCommandGroup extends SequentialCommandGroup {
+    public IntakeHandoffCommandGroup() {
         // Get arm into position
-        addCommands(new IntakeStowCommand());
+        addCommands(new IntakeArmStowCommand());
         addCommands(new RotateArmToPositionCommand(Constants.ARM_ROTATION.INTAKE_HANDOFF_ROTATIONS));
 
         // Grab the game piece
