+++ robotbase/src/main/java/com/team2357/frc2023/commands/intake/IntakeRollerReverseCommand.java
@@ -3,8 +3,8 @@ package com.team2357.frc2023.commands.intake;
 import com.team2357.frc2023.subsystems.IntakeRollerSubsystem;
 import com.team2357.lib.commands.CommandLoggerBase;
 
-public class IntakeReverseCommand extends CommandLoggerBase {
-    public IntakeReverseCommand() {
+public class IntakeRollerReverseCommand extends CommandLoggerBase {
+    public IntakeRollerReverseCommand() {
         addRequirements(IntakeRollerSubsystem.getInstance());
     }
