+++ robotbase/src/main/java/com/team2357/frc2023/commands/intake/IntakeStowCommandGroup.java
@@ -1,0 +1,7 @@
+package com.team2357.frc2023.commands.intake;
+
+import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+
+public class IntakeStowCommandGroup extends SequentialCommandGroup{
+    
+}
