+++ robotbase/src/main/java/com/team2357/frc2023/commands/scoring/AutoScoreHighCommand.java
@@ -1,7 +1,7 @@
 package com.team2357.frc2023.commands.scoring;
 
 import com.team2357.frc2023.Constants;
-import com.team2357.frc2023.commands.intake.IntakeReverseCommand;
+import com.team2357.frc2023.commands.intake.IntakeRollerReverseCommand;
 import com.team2357.frc2023.commands.scoring.util.ExtendArmToPositionCommand;
 import com.team2357.frc2023.commands.scoring.util.ExtendWristCommand;
 import com.team2357.frc2023.commands.scoring.util.OpenClawCommand;
@@ -15,7 +15,7 @@ public class AutoScoreHighCommand extends SequentialCommandGroup {
         // Pull game piece away from intake
         addCommands(new ParallelRaceGroup(
             new ExtendArmToPositionCommand(Constants.ARM_EXTENSION.RETRACTED_ROTATIONS),
-            new IntakeReverseCommand()
+            new IntakeRollerReverseCommand()
         ));
 
         // Extend to node
