+++ robotbase/src/main/java/com/team2357/frc2023/controls/ButtonBoardControls.java
@@ -1,7 +1,5 @@
 package com.team2357.frc2023.controls;
 
-import com.team2357.frc2023.Constants;
-
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
