+++ robotbase/src/main/java/com/team2357/frc2023/controls/GunnerControls.java
@@ -3,7 +3,6 @@ package com.team2357.frc2023.controls;
 import com.team2357.lib.triggers.AxisThresholdTrigger;
 import com.team2357.lib.util.Utility;
 import com.team2357.lib.util.XboxRaw;
-import com.team2357.frc2023.controls.AxisInterface;
 import com.team2357.frc2023.subsystems.SwerveDriveSubsystem;
 
 import edu.wpi.first.wpilibj.XboxController;
@@ -11,9 +10,6 @@ import edu.wpi.first.wpilibj.XboxController.Axis;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import edu.wpi.first.wpilibj2.command.button.POVButton;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
-import static com.team2357.frc2023.Constants.*;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.team2357.frc2023.Constants.CONTROLLER;
 import com.team2357.frc2023.commands.auto.TranslateToTargetCommand;
 import com.team2357.frc2023.commands.auto.TranslateToTargetCommandGroup;
